requirements.txt
google-genai
mss
Pillow
keyboard
pyttsx3
python-dotenv
